+++ CHANGELOG.rst
@@ -14,6 +14,25 @@ The format is based on `Keep a Changelog`_, and this project adheres to
 [Unreleased] - yyyy-mm-dd
 --------------------
 
+Added
+~~~~~
+* Published as filter.
+
+Changed
+~~~~~~~
+* Pseudonyms and 'published as anonymous' display in brackets after the author's name.
+* Index contributions by agent's name not pseudonyms.
+* Remove country names from the events classification filter.
+* Display related events in the events timeline.
+
+Removed
+~~~~~~~
+* Comparative value from the events classification filter.
+
+Fixed
+~~~~~
+* Display of duplicate contributions.
+
 
 [1.1.0] - 2021-05-17
 --------------------